Added generic AutocompleteItemDelegate for autocompletion in list views
